Move non-shadcn ui component to other folder
