+++ frontend/components/common/Diagram.tsx
