+++ frontend/components/common/tokenUsageIndicator.tsx
